Fix: Pytests. Change againt to .py extension
Fix: Improve exit codes
