+++ integrations/slack.py
@@ -10,19 +10,26 @@ import os
 import sys
 import time
 
+
+# Exit error codes
+ERR_NO_REQUEST_MODULE   = 1
+ERR_BAD_ARGUMENTS       = 2
+ERR_FILE_NOT_FOUND      = 6
+ERR_INVALID_JSON        = 7
+
 try:
     import requests
     from requests.auth import HTTPBasicAuth
 except Exception as e:
     print("No module 'requests' found. Install: pip install requests")
-    sys.exit(1)
+    sys.exit(ERR_NO_REQUEST_MODULE)
 
 # ossec.conf configuration structure
 #  <integration>
 #      <name>slack</name>
 #      <hook_url>https://hooks.slack.com/services/XXXXXXXXXXXXXX</hook_url>
 #      <alert_format>json</alert_format>
-#      <options>JSON_OBJ</options>
+#      <options>JSON</options> <!-- Replace with your custom JSON object -->
 #  </integration>
 
 # Global vars
@@ -66,7 +73,7 @@ def main(args: list[str]):
 
         if bad_arguments:
             debug("# Exiting: Bad arguments. Inputted: %s" % args)
-            sys.exit(2)
+            sys.exit(ERR_BAD_ARGUMENTS)
         
         # Core function
         process_args(args)
@@ -249,10 +256,10 @@ def get_json_alert(alert_file_location: str) -> any:
             return json.load(alert_file)
     except FileNotFoundError:
         debug("# Alert file %s doesn't exist" % alert_file_location)
-        sys.exit(3)
+        sys.exit(ERR_FILE_NOT_FOUND)
     except json.decoder.JSONDecodeError as e:
         debug("Failed getting json_alert %s" % e)
-        sys.exit(4)
+        sys.exit(ERR_INVALID_JSON)
         
 def get_json_options(options_file_location: str) -> any:
     """ 
@@ -280,10 +287,10 @@ def get_json_options(options_file_location: str) -> any:
             return json.load(options_file)
     except FileNotFoundError:
         debug("# Option file %s doesn't exist" % options_file_location)
-        sys.exit(3)
+        sys.exit(ERR_FILE_NOT_FOUND)
     except json.decoder.JSONDecodeError as e:
         debug("Failed getting json_alert %s" % e)
-        sys.exit(4)
+        sys.exit(ERR_INVALID_JSON)
 
 if __name__ == "__main__":
     main(sys.argv)
