+++ integrations/virustotal.py
@@ -5,25 +5,37 @@
 # License (version 2) as published by the FSF - Free Software
 # Foundation.
 
+
 import json
 import os
 import sys
 import time
+from socket import socket, AF_UNIX, SOCK_DGRAM
+
+# Exit error codes
+ERR_NO_REQUEST_MODULE   = 1
+ERR_BAD_ARGUMENTS       = 2
+ERR_BAD_MD5_SUM         = 3
+ERR_NO_RESPONSE_VT      = 4
+ERR_SOCKET_OPERATION    = 5
+ERR_FILE_NOT_FOUND      = 6
+ERR_INVALID_JSON        = 7
 
 try:
     import requests
     from requests.auth import HTTPBasicAuth
 except Exception as e:
     print("No module 'requests' found. Install: pip install requests")
-    sys.exit(1)
+    sys.exit(ERR_NO_REQUEST_MODULE)
 
-# ossec.conf configuration structure
-#  <integration>
-#      <name>slack</name>
-#      <hook_url>https://hooks.slack.com/services/XXXXXXXXXXXXXX</hook_url>
-#      <alert_format>json</alert_format>
-#      <options>JSON_OBJ</options>
-#  </integration>
+# ossec.conf configuration:
+# <integration>
+#   <name>virustotal</name>
+#   <api_key>API_KEY</api_key> <!-- Replace with your VirusTotal API key -->
+#   <group>syscheck</group>
+#   <alert_format>json</alert_format>
+#   <options>JSON</options> <!-- Replace with your custom JSON object -->
+# </integration>
 
 # Global vars
 debug_enabled   = False
@@ -33,8 +45,9 @@ json_alert      = {}
 json_options    = {}
 now             = time.strftime("%a %b %d %H:%M:%S %Z %Y")
 
-# Log path
+# Log and socket path
 LOG_FILE        = f'{pwd}/logs/integrations.log'
+SOCKET_ADDR     = f'{pwd}/queue/sockets/queue'
 
 # Constants
 ALERT_INDEX     = 1
@@ -66,14 +79,14 @@ def main(args: list[str]):
 
         if bad_arguments:
             debug("# Exiting: Bad arguments. Inputted: %s" % args)
-            sys.exit(2)
+            sys.exit(ERR_BAD_ARGUMENTS)
         
         # Core function
         process_args(args)
 
     except Exception as e:
         debug(str(e))
-        raise 
+        raise
     
 def process_args(args: list[str]) -> None:
     """ 
@@ -119,13 +132,12 @@ def process_args(args: list[str]) -> None:
     debug("# Generating message")
     msg: any = generate_msg(json_alert, json_options,apikey)
     
-    if not len(msg):
+    if not msg:
         debug("# ERR - Empty message")
         raise Exception
-    debug(msg)
-
+    
     debug("# Sending message")
-    send_msg(msg)
+    send_msg(msg,json_alert["agent"])
 
 def debug(msg: str) -> None:
     """ 
@@ -145,8 +157,7 @@ def debug(msg: str) -> None:
     if debug_console:
         msg = "{0}: {1}\n".format(now, msg)
         print(msg)
-
-
+        
 def generate_msg(alert: any, options: any,apikey:str) -> str:
     """ 
         Generate the JSON object with the message to be send
@@ -163,56 +174,117 @@ def generate_msg(alert: any, options: any,apikey:str) -> str:
         msg: str
             The JSON message to send
     """
-    managed_security_url    = 'https://wazuh.com'
-    level                   = alert['rule']['level']
-
-    severity = 'info'
-    if level >= 7:
-        severity = 'warning'
-    elif level >= 10:
-        severity = 'error'
-    elif level >= 13:
-        severity = 'critical'
+    msg = {}
+    # If there is no a md5 checksum present in the alert. Exit.
+    if not "md5_after" in alert["syscheck"]:
+        debug("# Exiting: MD5 checksum not found in alert.")
+        sys.exit(ERR_BAD_MD5_SUM)
     
-    groups = ', '.join(alert['rule']['groups'])
+    # Request info using VirusTotal API
+    try:
+        vt_response_data = query_api(alert["syscheck"]["md5_after"], apikey)
+    except Exception as e:
+        debug(e)
+        sys.exit(ERR_NO_RESPONSE_VT)
         
-    msg = {
-        'routing_key':  apikey,
-        'event_action': 'trigger',
-        'payload': {
-            "summary": alert['rule']['description'] if 'description' in alert['rule'] else "N/A",
-            "timestamp": alert['timestamp'],
-            "source": alert['agent']['location'],
-            "severity": severity,
-            "group": groups,
-            "custom_details": alert
-        },
-        "client": "Wazuh Monitoring Service",
-        "client_url": managed_security_url
-    }
-
+    msg["virustotal"]                           = {}
+    msg["integration"]                          = "virustotal"
+    msg["virustotal"]["found"]                  = 0
+    msg["virustotal"]["malicious"]              = 0
+    msg["virustotal"]["source"]                 = {}
+    msg["virustotal"]["source"]["alert_id"]     = alert["id"]
+    msg["virustotal"]["source"]["file"]         = alert["syscheck"]["path"]
+    msg["virustotal"]["source"]["md5"]          = alert["syscheck"]["md5_after"]
+    msg["virustotal"]["source"]["sha1"]         = alert["syscheck"]["sha1_after"]
+    
+    # Check if VirusTotal has any info about the hash
+    if vt_response_data['response_code']:
+        msg["virustotal"]["found"] = 1
+    
+    # Info about the file found in VirusTotal
+    if msg["virustotal"]["found"] == 1:
+        if vt_response_data['positives'] > 0:
+            msg["virustotal"]["malicious"] = 1
+        # Populate JSON Output object with VirusTotal request
+        msg["virustotal"]["sha1"]           = vt_response_data['sha1']
+        msg["virustotal"]["scan_date"]      = vt_response_data['scan_date']
+        msg["virustotal"]["positives"]      = vt_response_data['positives']
+        msg["virustotal"]["total"]          = vt_response_data['total']
+        msg["virustotal"]["permalink"]      = vt_response_data['permalink']
+        
     if(options):
         msg.update(options)
         
     return json.dumps(msg)
 
-def send_msg(msg: any) -> None:
+def query_api(hash: str, apikey: str) -> any:
     """ 
-        Send the message to the API
-
+        Send a request to VT API and fetch information to build message
+            
         Parameters
         ----------
-        msg : str
-            JSON message.
-        url: str
-            URL of the API.
+        hash : str
+            Hash need it for parameters
+        apikey: str
+            JSON options object.
+               
+        Returns
+        -------
+        data: any
+            JSON with the response
+            
+        Raises
+        ------
+        Exception
+            If the status code is different than 200.
     """
-    debug("# In send msg")
-    headers = {'content-type': 'application/json', 'Accept-Charset': 'UTF-8'}
-    url                = 'https://events.pagerduty.com/v2/enqueue'
-    res = requests.post(url, data=msg, headers=headers)
-    debug("# After send msg: %s" % res)
-    
+    params    = {'apikey': apikey, 'resource': hash}
+    headers   = {
+    "Accept-Encoding": "gzip, deflate",
+    "User-Agent" : "gzip,  Python library-client-VirusTotal"
+    }
+    response  = requests.get('https://www.virustotal.com/vtapi/v2/file/report',params=params, headers=headers)
+  
+    if response.status_code == 200:
+        json_response = response.json()
+        data = json_response
+        return data
+    else:
+        alert_output                  = {}
+        alert_output["virustotal"]    = {}
+        alert_output["integration"]   = "virustotal"
+
+        if response.status_code == 204:
+          alert_output["virustotal"]["error"]         = response.status_code
+          alert_output["virustotal"]["description"]   = "Error: Public API request rate limit reached"
+          raise Exception("# Error: VirusTotal Public API request rate limit reached")
+        elif response.status_code == 403:
+          alert_output["virustotal"]["error"]         = response.status_code
+          alert_output["virustotal"]["description"]   = "Error: Check credentials"
+          raise Exception("# Error: VirusTotal credentials, required privileges error")
+        else:
+          alert_output["virustotal"]["error"]         = response.status_code
+          alert_output["virustotal"]["description"]   = "Error: API request fail"
+          raise Exception("# Error: VirusTotal credentials, required privileges error")
+
+def send_msg(msg: any, agent:any) -> None:
+    if not agent or agent["id"] == "000":
+        string      = '1:virustotal:{0}'.format(msg)
+    else:
+        location    = '[{0}] ({1}) {2}'.format(agent["id"], agent["name"], agent["ip"] if "ip" in agent else "any")
+        location    = location.replace("|", "||").replace(":", "|:")
+        string      = '1:{0}->virustotal:{1}'.format(location,msg)
+
+    debug("# Final msg to send: %s" % string)
+    try:
+        sock = socket(AF_UNIX, SOCK_DGRAM)
+        sock.connect(SOCKET_ADDR)
+        sock.send(string.encode())
+        sock.close()
+    except FileNotFoundError:
+        debug(" # Error: Unable to open socket connection at %s" % SOCKET_ADDR)
+        sys.exit(ERR_SOCKET_OPERATION)
+
 def get_json_alert(alert_file_location: str) -> any:
     """ 
         Read the JSON object from alert file
@@ -239,10 +311,10 @@ def get_json_alert(alert_file_location: str) -> any:
             return json.load(alert_file)
     except FileNotFoundError:
         debug("# Alert file %s doesn't exist" % alert_file_location)
-        sys.exit(3)
+        sys.exit(ERR_FILE_NOT_FOUND)
     except json.decoder.JSONDecodeError as e:
         debug("Failed getting json_alert %s" % e)
-        sys.exit(4)
+        sys.exit(ERR_INVALID_JSON)
         
 def get_json_options(options_file_location: str) -> any:
     """ 
@@ -270,10 +342,10 @@ def get_json_options(options_file_location: str) -> any:
             return json.load(options_file)
     except FileNotFoundError:
         debug("# Option file %s doesn't exist" % options_file_location)
-        sys.exit(3)
+        sys.exit(ERR_FILE_NOT_FOUND)
     except json.decoder.JSONDecodeError as e:
         debug("Failed getting json_alert %s" % e)
-        sys.exit(4)
+        sys.exit(ERR_INVALID_JSON)
 
 if __name__ == "__main__":
     main(sys.argv)
